Adding documentation to Dashboard items
