+++ src/components/home/dashboard/ButtonText.tsx
@@ -1,5 +1,10 @@
+/**
+ * Simple button text component used by Dashboard buttons.
+ */
 import React from 'react';
 import { Text, TextProps } from 'react-native';
+
+// Styling
 import { DashboardStyles } from '../../../styles/home/dashboard.styles';
 
 const ButtonText: React.FC<TextProps> = ({ style, children, ...props }) => {
